Remove additional source files
'magic' is used for the 'file' command and should be included with
upstream version.

'newsl' are some MailScanner libraries. We may need to refer to these to
patch MailScanner, but they will not be needed in the final repo.

Upstream ESET should be compatible. Regardless, we don't need old Debian
dependencies.

Pyzor and Razor are packaged for CentOS