Create test runner
I'll be removing Perl::Critic (and possibly other unnecessary
dependencies), so the `just test` option in the Bootc repo will not
be able to simply run `prove` within the built image.

This new script will will install the required dependencies for testing
within the built image, then execute the tests in one script. This will
allow the final images to be smaller while still being tested.