Remove update_patch table
Patches will no longer be layered on top of the installed image, but
instead come from the Bootc container. There's no need to track this in
the DB and even if we wanted to, it would require parsing the data from
`bootc status` or similar. Instead, the Bootc image build should make
the version available in `/etc/os-release`.