Progress towards working DB
Both DBs are generated and initialized, but replication doesn't quite
seem to be there. It reports that it is set up, but IO is stuck at
'Connecting' and source DB position remains firmly ahead of replica.

Some fixes made to DB.pm to restore array and hashref fetches. Still
more testing/fixes to come.