Temporarily remove MailScanner dependency
I haven't finished building our version of MailScanner yet, so for now
disable the SystemD dependency so that Stages 2 and 1 can start.

I may potentially have Stage 2 (or 1) pass the message directly to
Stage 4 to test mail flow without filtering if I'm still working on
MailScanner when everything else is ready for testing.