Generate TLS cert during install
Fetches approximate location from ipinfo in order to create a config
file, then generates the key and cert and inserts them after the
database is up.

Defaults to Edmonton and SpamTagger as an org if IP cannot be found or
if not run by installer (if ORGANIZATION is not defined as an ENV var).